Do not reset bounding box when z drops
Prior to this change, the code would reset the bounding box
if the current z position was lowered and the previous z was
a printable area.

This means that if you have a z lift that hits a printable layer
then the bounding box will be reset.

This causes the bounding box to frequently collide with the print
when slicing with KISSlicer with a z lift and variable layer heights.